feat: 프로필 이미지 공통 컴포넌트 생성
- 스타일의 통일성과 매번 이미지 에러를 처리하지 않기 위해 생성
