+++ src/components/@common/ProfileImage/ProfileImage.styled.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const ProfileImg = styled.img<{ $size: number }>`
+  width: ${({ $size }) => `${$size}px`};
+  height: ${({ $size }) => `${$size}px`};
+  border-radius: 999px;
+  object-fit: cover;
+`;
