+++ src/components/@layout/SideBar/SideBarFriendList/index.tsx
@@ -1,6 +1,7 @@
 import useGetMyFriendList from '@/hooks/user/useGetMyFriendList';
 import SideBarListWrapper from '../SideBarListWrapper';
 import * as Styled from './SideBarFriendList.styled';
+import ProfileImage from '@/components/@common/ProfileImage';
 
 const SideBarFriendList = () => {
   const {
@@ -11,9 +12,10 @@ const SideBarFriendList = () => {
     <SideBarListWrapper isOpened title="친구" count={content.length}>
       {content.map((friend) => (
         <Styled.Wrapper key={friend.memberId}>
-          <Styled.ProfileImg
+          <ProfileImage
             src={friend.imageUrl}
             alt={`${friend.nickname} 프로필 이미지`}
+            size={56}
           />
           <Styled.FriendInfo>
             <Styled.Nickname>{friend.nickname}</Styled.Nickname>
