+++ src/components/@layout/SideBar/SideBarHeader/SideBarHeader.styled.ts
@@ -8,23 +8,6 @@ export const Wrapper = styled.section`
   padding: 0 24px;
 `;
 
-export const ProfileContent = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-`;
-
-export const ProfileImg = styled.img`
-  width: 52px;
-  height: 52px;
-  border-radius: 999px;
-`;
-
-export const Nickname = styled.p`
-  ${({ theme }) => theme.typo.T1_B}
-  color: ${({ theme }) => theme.colors.Gray_1000}
-`;
-
 export const NotificationBtn = styled.button`
   position: relative;
 `;
