+++ src/components/@layout/SideBar/SideBarProfile/index.tsx
@@ -1,17 +1,13 @@
 import { useProfile } from '@/hooks/user/useProfile';
-import * as Styled from '../SideBarHeader/SideBarHeader.styled';
+import * as Styled from './SideBarProfile.styled';
 import { useSetRecoilState } from 'recoil';
 import profileAtom from '@/recoil/user/profile';
-import { ReactEventHandler, useEffect } from 'react';
-import defaultProfileImg from '@/assets/images/defaultProfile.png';
+import { useEffect } from 'react';
+import ProfileImage from '@/components/@common/ProfileImage';
 
 const SideBarProfile = () => {
   const { data } = useProfile();
   const setProfile = useSetRecoilState(profileAtom);
-
-  const handleImgError: ReactEventHandler<HTMLImageElement> = (e) => {
-    e.currentTarget.src = defaultProfileImg;
-  };
 
   useEffect(() => {
     if (data) {
@@ -20,11 +16,7 @@ const SideBarProfile = () => {
   }, [data, setProfile]);
   return (
     <Styled.ProfileContent>
-      <Styled.ProfileImg
-        src={data.imageUrl}
-        alt="profile image"
-        onError={handleImgError}
-      />
+      <ProfileImage src={data.imageUrl} alt="profile image" size={52} />
       <Styled.Nickname>{data.nickname}</Styled.Nickname>
     </Styled.ProfileContent>
   );
